* Reduce argument passing cost. Change in funtion argument types to reduce runtime cost. Adding 'const' to reduce unwanted change risk. Only a trivial an uncomplete modification. We need better understanding of code to establish clear interfaces. Also clarify lifetime and ownership of dynamic storage.   * Simple warnings elimination.

+++ native/src/ElapsedTimer.cpp
@@ -40,7 +40,7 @@ const high_resolution_clock::duration& OsmAnd::ElapsedTimer::GetElapsed()
 	return elapsed;
 }
 
-uint64_t OsmAnd::ElapsedTimer::GetElapsedMs()
+int OsmAnd::ElapsedTimer::GetElapsedMs()
 {
     Pause();
     return std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count();
